Move files and namespaces to match MAUI format.
No functional changes, Simply move files, namespaces, rename variables, etc. so that the project structure more closely resembles the format that MAUI designers suggest. See: https://github.com/jsuarezruiz/xamarin-forms-to-net-maui/tree/main/Library/.NET%20MAUI/Library
