+++ Maui/HtmlLabel.Forms.Plugin.Tests.App/App.xaml
@@ -1,8 +1,8 @@
 ﻿<?xml version = "1.0" encoding = "UTF-8" ?>
 <Application xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:local="clr-namespace:HtmlLabel.Forms.Plugin.Tests.App"
-             x:Class="HtmlLabel.Forms.Plugin.Tests.App.App">
+             xmlns:local="clr-namespace:HyperTextLabel.Maui.Tests.App"
+             x:Class="HyperTextLabel.Maui.Tests.App.App">
     <Application.Resources>
         <ResourceDictionary Source="Resources/Styles.xaml" />
     </Application.Resources>
