+++ Maui/HtmlLabel.Forms.Plugin.Tests.App/AppShell.xaml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <Shell
-    x:Class="HtmlLabel.Forms.Plugin.Tests.App.AppShell"
+    x:Class="HyperTextLabel.Maui.Tests.App.AppShell"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:local="clr-namespace:HtmlLabel.Forms.Plugin.Tests.App"
+    xmlns:local="clr-namespace:HyperTextLabel.Maui.Tests.App"
     Shell.FlyoutBehavior="Disabled">
 
     <ShellContent
