+++ Maui/HtmlLabel.Forms.Plugin.Tests.App/HyperTextLabel.Maui.Tests.App.csproj
@@ -6,16 +6,16 @@
 		<!-- Uncomment to also build the tizen app. You will need to install tizen by following this: https://github.com/Samsung/Tizen.NET -->
 		<!-- <TargetFrameworks>$(TargetFrameworks);net6.0-tizen</TargetFrameworks> -->
 		<OutputType>Exe</OutputType>
-		<RootNamespace>HtmlLabel.Forms.Plugin.Tests.App</RootNamespace>
+		<RootNamespace>HyperTextLabel.Maui.Tests.App</RootNamespace>
 		<UseMaui>true</UseMaui>
 		<SingleProject>true</SingleProject>
 		<ImplicitUsings>enable</ImplicitUsings>
 
 		<!-- Display name -->
-		<ApplicationTitle>HtmlLabel.Forms.Plugin.Tests.App</ApplicationTitle>
+		<ApplicationTitle>HyperTextLabel.Maui.Tests.App</ApplicationTitle>
 
 		<!-- App Identifier -->
-		<ApplicationId>com.companyname.htmllabel.forms.plugin.tests.app</ApplicationId>
+		<ApplicationId>com.companyname.HyperTextLabel.Maui.Tests.App</ApplicationId>
 		<ApplicationIdGuid>F49B1950-D2D6-4246-861F-A061DA235263</ApplicationIdGuid>
 
 		<!-- Versions -->
@@ -50,7 +50,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <ProjectReference Include="..\HtmlLabel\HtmlLabel.Forms.Plugin.Maui.csproj" />
+	  <ProjectReference Include="..\HtmlLabel\HyperTextLabel.Maui.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
