+++ Maui/HtmlLabel.Forms.Plugin.Tests.App/MainPage.xaml
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:htmlLabel="clr-namespace:LabelHtml.Forms.Plugin;assembly=HtmlLabel.Forms.Plugin"
-             x:Class="HtmlLabel.Forms.Plugin.Tests.App.MainPage">
+             xmlns:htmlLabel="clr-namespace:HyperTextLabel.Maui.Controls;assembly=HyperTextLabel.Maui"
+             x:Class="HyperTextLabel.Maui.Tests.App.MainPage">
 
     <ScrollView>
         <StackLayout Padding="10">
