+++ Maui/HtmlLabel.Forms.Plugin.Tests.App/Platforms/Tizen/tizen-manifest.xml
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.companyname.HtmlLabel.Forms.Plugin.Tests.App" version="1.0.0" api-version="7" xmlns="http://tizen.org/ns/packages">
+<manifest package="com.companyname.HyperTextLabel.Maui.Tests.App" version="1.0.0" api-version="7" xmlns="http://tizen.org/ns/packages">
   <profile name="common" />
-  <ui-application appid="com.companyname.HtmlLabel.Forms.Plugin.Tests.App" exec="HtmlLabel.Forms.Plugin.Tests.App.dll" multiple="false" nodisplay="false" taskmanage="true" type="dotnet" launch_mode="single">
-    <label>HtmlLabel.Forms.Plugin.Tests.App</label>
+  <ui-application appid="com.companyname.HyperTextLabel.Maui.Tests.App" exec="HyperTextLabel.Maui.Tests.App.dll" multiple="false" nodisplay="false" taskmanage="true" type="dotnet" launch_mode="single">
+    <label>HyperTextLabel.Maui.Tests.App</label>
     <icon>appicon.xhigh.png</icon>
     <metadata key="http://tizen.org/metadata/prefer_dotnet_aot" value="true" />
   </ui-application>
