+++ Maui/HtmlLabel.Forms.Plugin.Tests.App/Platforms/Windows/App.xaml
@@ -1,8 +1,8 @@
 ﻿<maui:MauiWinUIApplication
-    x:Class="HtmlLabel.Forms.Plugin.Tests.App.WinUI.App"
+    x:Class="HyperTextLabel.Maui.Tests.App.WinUI.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:maui="using:Microsoft.Maui"
-    xmlns:local="using:HtmlLabel.Forms.Plugin.Tests.App.WinUI">
+    xmlns:local="using:HyperTextLabel.Maui.Tests.App.WinUI">
 
 </maui:MauiWinUIApplication>
