+++ Maui/HtmlLabel.Forms.Plugin.Tests.App/Platforms/Windows/App.xaml.cs
@@ -3,7 +3,7 @@
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
 
-namespace HtmlLabel.Forms.Plugin.Tests.App.WinUI;
+namespace HyperTextLabel.Maui.Tests.App.WinUI;
 
 /// <summary>
 /// Provides application-specific behavior to supplement the default Application class.
