+++ Maui/HtmlLabel.Forms.Plugin.Tests.App/Platforms/iOS/Info.plist
@@ -29,13 +29,13 @@
 	<key>MinimumOSVersion</key>
 	<string>14.2</string>
 	<key>CFBundleDisplayName</key>
-	<string>HtmlLabel.Forms.Plugin.Tests.App</string>
+	<string>HyperTextLabel.Maui.Tests.App</string>
 	<key>CFBundleIdentifier</key>
-	<string>com.companyname.HtmlLabel.Forms.Plugin.Tests.App</string>
+	<string>com.companyname.HyperTextLabel.Maui.Tests.App</string>
 	<key>CFBundleVersion</key>
 	<string>1.0</string>
 	<key>CFBundleName</key>
-	<string>HtmlLabel.Forms.Plugin.Tests.App</string>
+	<string>HyperTextLabel.Maui.Tests.App</string>
 	<key>LSApplicationQueriesSchemes</key>
 	<array>
 		<string>mailto</string>
