+++ Maui/HtmlLabel/HyperTextLabel.Maui.csproj
@@ -18,15 +18,15 @@
 	</PropertyGroup>
 
 	<PropertyGroup>
-		<AssemblyName>HtmlLabel.Forms.Plugin</AssemblyName>
-		<RootNamespace>LabelHtml.Forms.Plugin</RootNamespace>
-		<PackageId>Xam.Plugin.HtmlLabel</PackageId>
-		<Product>HtmlLabel.Forms.Plugin</Product>
+		<AssemblyName>HyperTextLabel.Maui</AssemblyName>
+		<RootNamespace>HyperTextLabel.Maui</RootNamespace>
+		<PackageId>Maui.Plugin.HtmlLabel</PackageId>
+		<Product>HtmlLabel.Maui.Plugin</Product>
 		<PackageIcon>icon.png</PackageIcon>
-		<Summary>Xam.Plugin.HtmlLabel: display HTML content in labels</Summary>
-		<PackageTags>xamarin, windows, uwp, ios, android, xamarin.forms, Xamarin.HtmlLabel, html</PackageTags>
-		<Title>Xam.Plugin.HtmlLabel</Title>
-		<Description>Xam.Plugin.HtmlLabel: display HTML content in labels</Description>
+		<Summary>Maui.Plugin.HtmlLabel: display HTML content in labels</Summary>
+		<PackageTags>MAUI, windows, uwp, ios, android, Maui.HtmlLabel, html</PackageTags>
+		<Title>Maui.Plugin.HtmlLabel</Title>
+		<Description>Maui.Plugin.HtmlLabel: display HTML content in labels</Description>
 		<Product>$(AssemblyName) ($(TargetFramework))</Product>
 		<PackageVersion>$(Version)$(VersionSuffix)</PackageVersion>
 		<Authors>Matteo Bortolazzo</Authors>
@@ -65,6 +65,7 @@
 	<ItemGroup>
 		<None Include="..\..\LICENSE" PackagePath="" Pack="true" />
 		<None Include="..\..\Assets\icon.png" PackagePath="" Pack="true" />
+		<None Include="Extensions\StringExtensions.cs" />
 		<PackageReference Include="Microsoft.CodeAnalysis.FxCopAnalyzers" Version="3.3.0">
 			<PrivateAssets>all</PrivateAssets>
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
