+++ Maui/HtmlLabel/Platforms/Windows/HtmlLabelExtensions.cs
@@ -1,44 +1,42 @@
 ﻿using Microsoft.UI.Xaml.Controls;
-using LabelHtml.Forms.Plugin.Abstractions;
-using LabelHtml.Forms.Plugin.UWP;
 using Microsoft.Xaml.Interactivity;
+using HyperTextLabel.Maui.Controls;
+using HyperTextLabel.Maui.Utilities;
 
-
-
-namespace LabelHtml.Forms.Plugin.Platforms.Windows
+namespace HyperTextLabel.Maui.Platforms.Windows
 {
     internal static class HtmlLabelExtensions
     {
-        public static void UpdateText(this TextBlock view, IHtmlLabel entry)
+        public static void UpdateText(this TextBlock view, IHtmlLabel label)
         {
-            ProcessText( view, entry);
+            ProcessText( view, label);
         }
 
-        public static void UpdateUnderlineText(this TextBlock view, IHtmlLabel entry)
-        {
-        }
-
-        public static void UpdateLinkColor(this TextBlock view, IHtmlLabel entry)
+        public static void UpdateUnderlineText(this TextBlock view, IHtmlLabel label)
         {
         }
 
-        public static void UpdateBrowserLaunchOptions(this TextBlock view, IHtmlLabel entry)
+        public static void UpdateLinkColor(this TextBlock view, IHtmlLabel label)
         {
         }
 
-        public static void UpdateAndroidLegacyMode(this TextBlock view, IHtmlLabel entry)
+        public static void UpdateBrowserLaunchOptions(this TextBlock view, IHtmlLabel label)
         {
         }
 
-        public static void UpdateAndroidListIndent(this TextBlock view, IHtmlLabel entry)
+        public static void UpdateAndroidLegacyMode(this TextBlock view, IHtmlLabel label)
         {
         }
 
-        private static void ProcessText(TextBlock view, IHtmlLabel entry)
+        public static void UpdateAndroidListIndent(this TextBlock view, IHtmlLabel label)
+        {
+        }
+
+        private static void ProcessText(TextBlock view, IHtmlLabel label)
         {
             // Gets the complete HTML string
             var isRtl = AppInfo.RequestedLayoutDirection == LayoutDirection.RightToLeft;
-            var styledHtml = new RendererHelper(entry, entry.Text, DevicePlatform.WinUI, isRtl).ToString();
+            var styledHtml = new RendererHelper(label, label.Text, DevicePlatform.WinUI, isRtl).ToString();
             if (styledHtml == null)
             {
                 return;
@@ -48,7 +46,7 @@ namespace LabelHtml.Forms.Plugin.Platforms.Windows
 
             // Adds the HtmlTextBehavior because UWP's TextBlock
             // does not natively support HTML content
-            var behavior = new HtmlTextBehavior() { HtmlLabel = entry };
+            var behavior = new HtmlTextBehavior() { HtmlLabel = label };
             BehaviorCollection behaviors = Interaction.GetBehaviors(view);
             behaviors.Clear();
             behaviors.Add(behavior);
