+++ Maui/HtmlLabel/Platforms/iOS/HtmlLabelHandler.cs
@@ -1,51 +1,54 @@
 ﻿using Microsoft.Maui.Platform;
-using LabelHtml.Forms.Plugin.Abstractions;
-using LabelHtml.Forms.Plugin.Platforms.iOS;
+using HyperTextLabel.Maui.Controls;
+using HyperTextLabel.Maui.Extensions;
+using HyperTextLabel.Maui.Platforms.iOS;
 
-namespace LabelHtml.Forms.Plugin
+using PlatformView = Microsoft.Maui.Platform.MauiLabel;
+
+namespace HyperTextLabel.Maui.Handlers
 {
     public partial class HtmlLabelHandler : Microsoft.Maui.Handlers.LabelHandler
     {
-        protected override void ConnectHandler(MauiLabel nativeView)
+        protected override void ConnectHandler(PlatformView platformView)
         {
-            base.ConnectHandler(nativeView);
+            base.ConnectHandler(platformView);
         }
 
-        protected override void DisconnectHandler(MauiLabel nativeView)
+        protected override void DisconnectHandler(PlatformView platformView)
         {
-            base.DisconnectHandler(nativeView);
+            base.DisconnectHandler(platformView);
         }
 
-        public static void MapLabelText(HtmlLabelHandler handler, IHtmlLabel entry)
+        public static void MapLabelText(HtmlLabelHandler handler, IHtmlLabel label)
         {
             var fontManager = handler.GetRequiredService<IFontManager>();
 
-            handler.PlatformView?.UpdateText(entry, fontManager);
+            handler.PlatformView?.UpdateText(label, fontManager);
         }
 
-        public static void MapUnderlineText(HtmlLabelHandler handler, IHtmlLabel entry)
+        public static void MapUnderlineText(HtmlLabelHandler handler, IHtmlLabel label)
         {
-            handler.PlatformView?.UpdateUnderlineText(entry);
+            handler.PlatformView?.UpdateUnderlineText(label);
         }
 
-        public static void MapLinkColor(HtmlLabelHandler handler, IHtmlLabel entry)
+        public static void MapLinkColor(HtmlLabelHandler handler, IHtmlLabel label)
         {
-            handler.PlatformView?.UpdateLinkColor(entry);
+            handler.PlatformView?.UpdateLinkColor(label);
         }
 
-        public static void MapBrowserLaunchOptions(HtmlLabelHandler handler, IHtmlLabel entry)
+        public static void MapBrowserLaunchOptions(HtmlLabelHandler handler, IHtmlLabel label)
         {
-            handler.PlatformView?.UpdateBrowserLaunchOptions(entry);
+            handler.PlatformView?.UpdateBrowserLaunchOptions(label);
         }
 
-        public static void MapAndroidLegacyMode(HtmlLabelHandler handler, IHtmlLabel entry)
+        public static void MapAndroidLegacyMode(HtmlLabelHandler handler, IHtmlLabel label)
         {
-            handler.PlatformView?.UpdateAndroidLegacyMode(entry);
+            handler.PlatformView?.UpdateAndroidLegacyMode(label);
         }
 
-        public static void MapAndroidListIndent(HtmlLabelHandler handler, IHtmlLabel entry)
+        public static void MapAndroidListIndent(HtmlLabelHandler handler, IHtmlLabel label)
         {
-            handler.PlatformView?.UpdateAndroidListIndent(entry);
+            handler.PlatformView?.UpdateAndroidListIndent(label);
         }
     }
 }
